add bk7271 support with 3.1.6 base

+++ beken378/ip/lmac/src/sta/sta_mgmt.h
@@ -283,7 +283,7 @@ struct sta_info_tag
     /// TX policy table control
     struct sta_pol_tbl_cntl pol_tbl;
 	
-	uint32_t pre_rx_timepoint;
+	uint64_t pre_rx_timepoint;
 
     #if (NX_AMPDU_TX)
     /// For TX BA agreements per TID
